Fix ForwardDiff Float64(::Dual) regression + round-7/8 audit bugs
CRITICAL regression fix (cause of the 2de0268 CI failure): the round-5/6
finiteness guards wrapped values in Float64(...), but Float64(::ForwardDiff.
Dual) is undefined, so texture sampling and soft_render crashed (MethodError)
the moment they ran through ForwardDiff — breaking the differentiable/inverse
rendering feature and an existing differentiable test. isfinite / comparisons /
clamp work directly on Float64, ForwardDiff.Dual, AND ADVar, so the Float64()
wrapper is dropped in:
- textures: _sanitize_uv, sample_texture_auto/aniso footprint guards
- soft_rasterizer: soft_render bbox finiteness guard

Round 7 (each with a regression test):
- math: mat4_perspective far=Inf gives a finite projection (three.js parity),
  not a NaN depth row
- controls: NaN query time -> clean ArgumentError across interpolate_catmull_rom
  / sample_track / mixer_set_time! (was a bare BoundsError)

Round 8:
- reverse_ad: ^(::ADVar, ::Rational) overload resolves a dispatch ambiguity
  with Base's ^(::Number, ::Rational)

Known remaining (deferred): transparent meshes render flat instead of smooth
(pre-existing render! classification gap; not a crash).

+++ src/soft_rasterizer.jl
@@ -104,9 +104,12 @@ function soft_render(vertices::Vector{Vec3{Tv}},
         # finite screen coordinate cannot overflow floor/ceil(Int, …) — matching
         # the hard rasterizer. Reachable in inverse rendering when an optimized
         # vertex transiently slides through the camera.
-        finite = isfinite(Float64(s1.x)) && isfinite(Float64(s1.y)) &&
-                 isfinite(Float64(s2.x)) && isfinite(Float64(s2.y)) &&
-                 isfinite(Float64(s3.x)) && isfinite(Float64(s3.y))
+        # isfinite works directly on Float64, ForwardDiff.Dual, and ADVar; do NOT
+        # wrap in Float64() — Float64(::ForwardDiff.Dual) is undefined and would
+        # crash the differentiable/inverse-rendering gradient path (its core use).
+        finite = isfinite(s1.x) && isfinite(s1.y) &&
+                 isfinite(s2.x) && isfinite(s2.y) &&
+                 isfinite(s3.x) && isfinite(s3.y)
         fW = Float64(W); fH = Float64(H)
         if finite
             bmin_x = max(floor(Int, clamp(min(s1.x, s2.x, s3.x) - margin, 1.0, fW)), 1)
